feat: Dealer 클래스에서의 ace 처리하는 로직 구현
