+++ README.md
@@ -18,10 +18,13 @@
 - [x] 두명의 점수를 비교하여 승패를 결정한다.
 
 ### BlackJackReferee
-- [ ] 최종 승패를 결정한다.
+- [x] 최종 승패를 결정한다.
 
 ### Player
 - [ ] 플레이어는 들고있는 카드의 총합이 21이하라면 카드를 추가로 받을 수 있다.
+- [ ] 더 뽑을거냐고 물어볼때는 Ace 무조건 1로 계산
 
 ### Dealer
 - [x] 딜러는 들고있는 카드의 총합이 16이하라면 카드를 추가로 받는다.
+- [x] 가지고 있는 카드 중 Ace 카드가 한장인 경우 무조건 11로 계산
+- [x] 가지고 있는 Ace 카드가 2장 이상이면 한장만 11, 나머지는 1로 점수 계산
