+++ src/main/kotlin/CardNumber.kt
@@ -1,4 +1,4 @@
 enum class CardNumber(val value: Int) {
-    A(0), ONE(1), TWO(2), THREE(3), FOUR(4), FIVE(5),
+    A(1), ONE(1), TWO(2), THREE(3), FOUR(4), FIVE(5),
     SIX(6), SEVEN(7), EIGHT(8), NINE(9), TEN(10), J(10), Q(10), K(10)
 }
