+++ src/main/kotlin/Dealer.kt
@@ -1,8 +1,27 @@
 class Dealer(val cards: Cards) {
 
+    fun isPossibleToDraw(score: Int): Boolean {
+        if (cards.cards.sumOf { card -> card.number.value } >= score)
+            return false
+
+        return true
+    }
+
     fun drawCard() {
-        if (cards.isPossibleToDraw(17)) {
+        if (isPossibleToDraw(17)) {
             cards.draw()
         }
+    }
+
+    fun getSum(): Int {
+        val count = cards.cards.count { card -> card.number == CardNumber.A }
+        var sum = cards.cards.sumOf { card -> card.number.value }
+        repeat(count) {
+            sum += 10
+            if (sum > 21)
+                return sum - 10
+        }
+
+        return sum
     }
 }
